// packages/clerk/src/operations/platform/users/banUser.ts
export const BanUserInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct({
  applicationID: Schema.String.pipe(T.PathParam()),
  envOrInsID: Schema.String.pipe(T.PathParam()),
  userID: Schema.String.pipe(T.PathParam()),
}).pipe(
  T.Http({
    method: "POST",
    path: "/platform/applications/{applicationID}/instances/{envOrInsID}/users/{userID}/ban",
  }),
);
export type BanUserInput = typeof BanUserInput.Type;
// ...
export type BanUserOutput = typeof BanUserOutput.Type;
// ...
export const banUser = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: BanUserInput,
  outputSchema: BanUserOutput,
  errors: [BadRequest, Forbidden, NotFound] as const,
}));