// packages/clerk/src/operations/platform/users/listInstanceUsers.ts
export const ListInstanceUsersInput = /*@__PURE__*/ /*#__PURE__*/ Schema.Struct(
  {
    applicationID: Schema.String.pipe(T.PathParam()),
    envOrInsID: Schema.String.pipe(T.PathParam()),
    query: Schema.optional(Schema.String),
    order_by: Schema.optional(Schema.String),
    limit: Schema.optional(Schema.Number),
    offset: Schema.optional(Schema.Number),
  },
).pipe(
  T.Http({
    method: "GET",
    path: "/platform/applications/{applicationID}/instances/{envOrInsID}/users",
  }),
);
export type ListInstanceUsersInput = typeof ListInstanceUsersInput.Type;
// ...
export type ListInstanceUsersOutput = typeof ListInstanceUsersOutput.Type;
// ...
export const listInstanceUsers = /*@__PURE__*/ /*#__PURE__*/ API.make(() => ({
  inputSchema: ListInstanceUsersInput,
  outputSchema: ListInstanceUsersOutput,
  errors: [BadRequest, Forbidden, NotFound, UnprocessableEntity] as const,
}));